+++ pkg/compactor/block_visit_marker.go
@@ -16,9 +16,16 @@ import (
 	"github.com/prometheus/client_golang/prometheus"
 	"github.com/thanos-io/objstore"
 	"github.com/thanos-io/thanos/pkg/block/metadata"
+
+	"github.com/cortexproject/cortex/pkg/util/runutil"
 )
 
-const BlockVisitMarkerFile = "block.visit"
+const (
+	// BlockVisitMarkerFile is the known json filename for representing the most recent compactor visit.
+	BlockVisitMarkerFile = "visit-mark.json"
+	// VisitMarkerVersion1 is the current supported version of visit-mark file.
+	VisitMarkerVersion1 = 1
+)
 
 var (
 	ErrorBlockVisitMarkerNotFound  = errors.New("block visit marker not found")
@@ -26,21 +33,24 @@ var (
 )
 
 type BlockVisitMarker struct {
-	CompactorID string    `json:"compactorID"`
-	VisitTime   time.Time `json:"visitTime"`
+	CompactorID string `json:"compactorID"`
+	// VisitTime is a unix timestamp of when the block was visited (mark updated).
+	VisitTime int64 `json:"visitTime"`
+	// Version of the file.
+	Version int `json:"version"`
 }
 
 func (b *BlockVisitMarker) isVisited(blockVisitMarkerTimeout time.Duration) bool {
-	return time.Now().Before(b.VisitTime.Add(blockVisitMarkerTimeout))
+	return time.Now().Before(time.Unix(b.VisitTime, 0).Add(blockVisitMarkerTimeout))
 }
 
 func (b *BlockVisitMarker) isVisitedByCompactor(blockVisitMarkerTimeout time.Duration, compactorID string) bool {
-	return time.Now().Before(b.VisitTime.Add(blockVisitMarkerTimeout)) && b.CompactorID == compactorID
+	return b.CompactorID == compactorID && time.Now().Before(time.Unix(b.VisitTime, 0).Add(blockVisitMarkerTimeout))
 }
 
-func ReadBlockVisitMarker(ctx context.Context, bkt objstore.Bucket, blockID string, blockVisitMarkerReadFailed prometheus.Counter) (*BlockVisitMarker, error) {
+func ReadBlockVisitMarker(ctx context.Context, bkt objstore.InstrumentedBucketReader, logger log.Logger, blockID string, blockVisitMarkerReadFailed prometheus.Counter) (*BlockVisitMarker, error) {
 	visitMarkerFile := path.Join(blockID, BlockVisitMarkerFile)
-	visitMarkerFileReader, err := bkt.Get(ctx, visitMarkerFile)
+	visitMarkerFileReader, err := bkt.ReaderWithExpectedErrs(bkt.IsObjNotFoundErr).Get(ctx, visitMarkerFile)
 	if err != nil {
 		if bkt.IsObjNotFoundErr(err) {
 			return nil, errors.Wrapf(ErrorBlockVisitMarkerNotFound, "block visit marker file: %s", visitMarkerFile)
@@ -48,46 +58,52 @@ func ReadBlockVisitMarker(ctx context.Context, bkt objstore.Bucket, blockID stri
 		blockVisitMarkerReadFailed.Inc()
 		return nil, errors.Wrapf(err, "get block visit marker file: %s", visitMarkerFile)
 	}
+	defer runutil.CloseWithLogOnErr(logger, visitMarkerFileReader, "close block visit marker reader")
 	b, err := io.ReadAll(visitMarkerFileReader)
 	if err != nil {
 		blockVisitMarkerReadFailed.Inc()
 		return nil, errors.Wrapf(err, "read block visit marker file: %s", visitMarkerFile)
 	}
 	blockVisitMarker := BlockVisitMarker{}
-	err = json.Unmarshal(b, &blockVisitMarker)
-	if err != nil {
+	if err = json.Unmarshal(b, &blockVisitMarker); err != nil {
 		blockVisitMarkerReadFailed.Inc()
 		return nil, errors.Wrapf(ErrorUnmarshalBlockVisitMarker, "block visit marker file: %s, error: %v", visitMarkerFile, err.Error())
+	}
+	if blockVisitMarker.Version != VisitMarkerVersion1 {
+		return nil, errors.Errorf("unexpected block visit mark file version %d, expected %d", blockVisitMarker.Version, VisitMarkerVersion1)
 	}
 	return &blockVisitMarker, nil
 }
 
-func UpdateBlockVisitMarker(ctx context.Context, bkt objstore.Bucket, blockID string, compactorID string, blockVisitMarkerWriteFailed prometheus.Counter) error {
+func UpdateBlockVisitMarker(ctx context.Context, bkt objstore.Bucket, blockID string, reader io.Reader, blockVisitMarkerWriteFailed prometheus.Counter) error {
 	blockVisitMarkerFilePath := path.Join(blockID, BlockVisitMarkerFile)
-	blockVisitMarker := BlockVisitMarker{
-		CompactorID: compactorID,
-		VisitTime:   time.Now(),
-	}
-	visitMarkerFileContent, err := json.Marshal(blockVisitMarker)
-	if err != nil {
-		blockVisitMarkerWriteFailed.Inc()
-		return err
-	}
-	err = bkt.Upload(ctx, blockVisitMarkerFilePath, bytes.NewReader(visitMarkerFileContent))
-	if err != nil {
+	if err := bkt.Upload(ctx, blockVisitMarkerFilePath, reader); err != nil {
 		blockVisitMarkerWriteFailed.Inc()
 		return err
 	}
 	return nil
 }
 
-func markBlocksVisited(ctx context.Context, bkt objstore.Bucket, logger log.Logger, blocks []*metadata.Meta, compactorID string, blockVisitMarkerWriteFailed prometheus.Counter) {
+func markBlocksVisited(
+	ctx context.Context,
+	bkt objstore.Bucket,
+	logger log.Logger,
+	blocks []*metadata.Meta,
+	marker BlockVisitMarker,
+	blockVisitMarkerWriteFailed prometheus.Counter,
+) {
+	visitMarkerFileContent, err := json.Marshal(marker)
+	if err != nil {
+		blockVisitMarkerWriteFailed.Inc()
+		return
+	}
+	reader := bytes.NewReader(visitMarkerFileContent)
 	for _, block := range blocks {
 		blockID := block.ULID.String()
-		err := UpdateBlockVisitMarker(ctx, bkt, blockID, compactorID, blockVisitMarkerWriteFailed)
-		if err != nil {
+		if err := UpdateBlockVisitMarker(ctx, bkt, blockID, reader, blockVisitMarkerWriteFailed); err != nil {
 			level.Error(logger).Log("msg", "unable to upsert visit marker file content for block", "blockID", blockID, "err", err)
 		}
+		reader.Reset(visitMarkerFileContent)
 	}
 }
 
@@ -103,7 +119,12 @@ func markBlocksVisitedHeartBeat(ctx context.Context, bkt objstore.Bucket, logger
 heartBeat:
 	for {
 		level.Debug(logger).Log("msg", fmt.Sprintf("heart beat for blocks: %s", blocksInfo))
-		markBlocksVisited(ctx, bkt, logger, blocks, compactorID, blockVisitMarkerWriteFailed)
+		blockVisitMarker := BlockVisitMarker{
+			VisitTime:   time.Now().Unix(),
+			CompactorID: compactorID,
+			Version:     VisitMarkerVersion1,
+		}
+		markBlocksVisited(ctx, bkt, logger, blocks, blockVisitMarker, blockVisitMarkerWriteFailed)
 
 		select {
 		case <-ctx.Done():