+++ pkg/compactor/shuffle_sharding_planner_test.go
@@ -16,6 +16,7 @@ import (
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/mock"
 	"github.com/stretchr/testify/require"
+	"github.com/thanos-io/objstore"
 	"github.com/thanos-io/thanos/pkg/block/metadata"
 
 	"github.com/cortexproject/cortex/pkg/storage/bucket"
@@ -348,7 +349,8 @@ func TestShuffleShardingPlanner_Plan(t *testing.T) {
 				}
 				blockVisitMarker := BlockVisitMarker{
 					CompactorID: visitedBlock.compactorID,
-					VisitTime:   expireTime,
+					VisitTime:   expireTime.Unix(),
+					Version:     VisitMarkerVersion1,
 				}
 				visitMarkerFileContent, _ := json.Marshal(blockVisitMarker)
 				bkt.MockGet(visitMarkerFile, string(visitMarkerFileContent), nil)
@@ -369,7 +371,7 @@ func TestShuffleShardingPlanner_Plan(t *testing.T) {
 			logger := log.NewLogfmtLogger(logs)
 			p := NewShuffleShardingPlanner(
 				context.Background(),
-				bkt,
+				objstore.WithNoopInstr(bkt),
 				logger,
 				testData.ranges,
 				func() map[ulid.ULID]*metadata.NoCompactMark {