+++ pkg/querier/tripperware/instantquery/shard_by_query_test.go
@@ -1,0 +1,12 @@
+package instantquery
+
+import (
+	"testing"
+
+	"github.com/cortexproject/cortex/pkg/querier/tripperware"
+	"github.com/cortexproject/cortex/pkg/querier/tripperware/queryrange"
+)
+
+func Test_shardQuery(t *testing.T) {
+	tripperware.TestQueryShardQuery(t, InstantQueryCodec, queryrange.ShardedPrometheusCodec)
+}