+++ pkg/querier/tripperware/queryrange/split_by_interval.go
@@ -69,6 +69,11 @@ func (s splitByInterval) Do(ctx context.Context, r tripperware.Request) (tripper
 }
 
 func splitQuery(r tripperware.Request, interval time.Duration) ([]tripperware.Request, error) {
+	// If Start == end we should just run the original request
+	if r.GetStart() == r.GetEnd() {
+		return []tripperware.Request{r}, nil
+	}
+
 	// Replace @ modifier function to their respective constant values in the query.
 	// This way subqueries will be evaluated at the same time as the parent query.
 	query, err := evaluateAtModifierFunction(r.GetQuery(), r.GetStart(), r.GetEnd())